Basic packet parsing done
It needs a bit of optimization, primarily around the regex (compile once). In a
quick test this processed ~1M packets at a rate of ~52k packets per second while
dropping 11 of them.

With a bit of clean I think we can get that number to 0. It's going to be
important to hand these stats off to a thread so that capture doesn't slow down.