Support for capturing loopback packets
This commit is a small refactor of the MemcacheCommand class. The interface
changed but the important part is that the optimization that detects whether the
packet is a request or a response (and ignores requests) has been pulled out. I
think this can be added back but the MemcacheCommand class likely needs a handle
on the device to check if the capture is occurring on a loopback or not.